[refactor] add decoratedExecutor and executorFactory
