+++ TestShell/executor.cpp
@@ -1,159 +1,12 @@
-#include <fstream>
-#include <sstream>
-#include <direct.h>
-#include <tuple>
-#include <algorithm>
 #include "executor.h"
 #include "compositeExecutor.h"
 #include "utils.h"
 #include "logger.h"
 
-using namespace std;
-
-const unordered_map<string, CompositeExecutorFactory::Creator> CompositeExecutorFactory::factoryMap = {
-	{FIRST_SCRIPT_SHORT_NAME,  []() { return new FullWriteAndReadCompare(new Writer(), new Comparer()); }},
-	{FIRST_SCRIPT_FULL_NAME,   []() { return new FullWriteAndReadCompare(new Writer(), new Comparer()); }},
-	{SECOND_SCRIPT_SHORT_NAME, []() { return new PartialLBAWrite(new Writer(), new Comparer()); }},
-	{SECOND_SCRIPT_FULL_NAME,  []() { return new PartialLBAWrite(new Writer(), new Comparer()); }},
-	{THIRD_SCRIPT_SHORT_NAME,  []() { return new WriteReadAging(new Writer(), new Comparer()); }},
-	{THIRD_SCRIPT_FULL_NAME,   []() { return new WriteReadAging(new Writer(), new Comparer()); }},
-	{FOURTH_SCRIPT_SHORT_NAME, []() { return new EraseAndWriteAging(new Writer(), new Comparer(), new Eraser()); }},
-	{FOURTH_SCRIPT_FULL_NAME,  []() { return new EraseAndWriteAging(new Writer(), new Comparer(), new Eraser()); }},
-};
-
-IExecutor* IExecutorFactory::getExecutorFromCache(const string& command) {
-	auto cacheIterator = executorCache.find(command);
-	if (cacheIterator != executorCache.end()) {
-		return cacheIterator->second.get();
-	}
-
-	return nullptr;
-}
-
-IExecutor* CompositeExecutorFactory::createExecutor(const string& command) {
-	IExecutor* executor = getExecutorFromCache(command);
-	if (nullptr != executor) return executor;
-
-	executor = getExecutorFromMap(command);
-	if (nullptr != executor) return executor;
-	
-	return nullptr;
-}
-
-IExecutor* CompositeExecutorFactory::getExecutorFromMap(const string& command) {
-	auto mapIterator = factoryMap.find(command);
-	if (mapIterator != factoryMap.end()) {
-		std::unique_ptr<IExecutor> executor(mapIterator->second());
-		IExecutor* rawPtr = executor.get();
-		executorCache[command] = std::move(executor);
-		return rawPtr;
-	}
-
-	return nullptr;
-}
-
-const unordered_map<string, ExecutorFactory::Creator> ExecutorFactory::factoryMap = {
-	{WRITE_CMD,                []() { return new OuputDecoratedWriter(); }},
-	{FULL_WRITE_CMD,           []() { return new FullWriter(); }},
-	{READ_CMD,                 []() { return new OuputDecoratedReader(); }},
-	{FULL_READ_CMD,            []() { return new FullReader(); }},
-	{ERASE_CMD,                []() { return new SizeEraser(); }},
-	{ERASE_RANGE_CMD,          []() { return new RangeEraser(); }},
-	{HELP_CMD,                 []() { return new Helper(); }},
-	{EXIT_CMD,                 []() { return new Exiter(); }},
-	{FLUSH_CMD,                []() { return new Flusher(); }},
-
-	{FIRST_SCRIPT_SHORT_NAME,  []() { return new FullWriteAndReadCompare(new Writer(), new Comparer()); }},
-	{FIRST_SCRIPT_FULL_NAME,   []() { return new FullWriteAndReadCompare(new Writer(), new Comparer()); }},
-	{SECOND_SCRIPT_SHORT_NAME, []() { return new PartialLBAWrite(new Writer(), new Comparer()); }},
-	{SECOND_SCRIPT_FULL_NAME,  []() { return new PartialLBAWrite(new Writer(), new Comparer()); }},
-	{THIRD_SCRIPT_SHORT_NAME,  []() { return new WriteReadAging(new Writer(), new Comparer()); }},
-	{THIRD_SCRIPT_FULL_NAME,   []() { return new WriteReadAging(new Writer(), new Comparer()); }},
-	{FOURTH_SCRIPT_SHORT_NAME, []() { return new EraseAndWriteAging(new Writer(), new Comparer(), new Eraser()); }},
-	{FOURTH_SCRIPT_FULL_NAME,  []() { return new EraseAndWriteAging(new Writer(), new Comparer(), new Eraser()); }},
-};
-
-IExecutor* ExecutorFactory::createExecutor(const string& command) {
-	IExecutor* executor = getExecutorFromCache(command);
-	if (nullptr != executor) return executor;
-
-	executor = getExecutorFromMap(command);
-	if (nullptr != executor) return executor;
-
-	return nullptr;
-}
-
-IExecutor* ExecutorFactory::getExecutorFromMap(const string& command) {
-	auto mapIterator = factoryMap.find(command);
-	if (mapIterator != factoryMap.end()) {
-		std::unique_ptr<IExecutor> executor(mapIterator->second());
-		IExecutor* rawPtr = executor.get();
-		executorCache[command] = std::move(executor);
-		return rawPtr;
-	}
-
-	return nullptr;
-}
-
 bool Writer::execute(ISsdApp* app, LBA lba, DATA data)
 {
 	app->Write(lba, data);
 	SHELL_LOG("[Write] Done", lba, data);
-	return true;
-}
-
-bool OuputDecoratedWriter::IsValidCommand(const vector<string>& tokens) {
-	if (tokens[CMD_IDX] == WRITE_CMD && tokens.size() == NEEDED_TOKEN_COUNT) {
-		if (IsValidWriteData(tokens[DATA_IDX])) return true;
-		return false;
-	}
-
-	dbgPrint("Command parameter is not enough! needed param count:" + to_string(NEEDED_TOKEN_COUNT) + ", current param count:" + to_string(tokens.size()));
-	return false;
-}
-
-bool OuputDecoratedWriter::IsValidWriteData(const string& data) {
-	if (data.size() != WRITE_DATA_SIZE) {
-		print("Write Data Invalid: 10글자가 안 됨 - DATA:" + data);
-		return false;
-	}
-
-	if (!(data[0] == '0' && (data[1] == 'x' || data[1] == 'X'))) {
-		print("Write Data Invalid: 0x로 시작하지 않음 - DATA:" + data);
-		return false;
-	}
-
-	for (size_t i = START_DATA_DIGIT_EXCEPT_HEX_PREFIX; i < WRITE_DATA_SIZE; ++i) {
-		if (!std::isxdigit(static_cast<unsigned char>(data[i]))) {
-			print("Write Data Invalid: 숫자가 아닌 값이 들어가 있음 - DATA:" + data);
-			return false;
-		}
-	}
-
-	return true;
-}
-
-bool OuputDecoratedWriter::execute(ISsdApp* app, LBA lba, DATA data)
-{
-	if (!Writer::execute(app, lba, data)) return false;
-	print("[Write] Done");
-	return true;
-}
-
-bool FullWriter::IsValidCommand(const vector<string>& tokens) {
-	if (tokens[CMD_IDX] == CMD && tokens.size() == NEEDED_TOKEN_COUNT) return true;
-	dbgPrint("Command parameter is not enough! needed param count:" + to_string(NEEDED_TOKEN_COUNT) + ", current param count:" + to_string(tokens.size()));
-	return false;
-}
-
-bool FullWriter::execute(ISsdApp* app, LBA lba, DATA data)
-{
-	DATA fullWriteData = lba;
-	for (LBA lba = 0; lba < SSD_MAX_SIZE; ++lba)
-	{
-		if (!Writer::execute(app, lba, fullWriteData)) return false;
-	}
-	print("[Full Write] Done");
 	return true;
 }
 
@@ -179,55 +32,6 @@ string Reader::GetResultFromFile(void)
 
 	file.close();
 	return result;
-}
-
-bool OuputDecoratedReader::IsValidCommand(const vector<string>& tokens) {
-	if (tokens[CMD_IDX] == CMD && tokens.size() == NEEDED_TOKEN_COUNT) return true;
-	dbgPrint("Command parameter is not enough! needed param count:" + to_string(NEEDED_TOKEN_COUNT) + ", current param count:" + to_string(tokens.size()));
-	return false;
-}
-
-bool OuputDecoratedReader::execute(ISsdApp* app, LBA lba, DATA data)
-{
-	if (!Reader::execute(app, lba, data)) return false;
-	std::ostringstream oss;
-	oss << "[Read] LBA:" << setw(2) << lba << " : " << GetResultFromFile();
-	print(oss.str());
-	return true;
-}
-
-bool FullReader::IsValidCommand(const vector<string>& tokens) {
-	if (tokens[CMD_IDX] == CMD && tokens.size() == NEEDED_TOKEN_COUNT) return true;
-	dbgPrint("Command parameter is not enough! needed param count:" + to_string(NEEDED_TOKEN_COUNT) + ", current param count:" + to_string(tokens.size()));
-	return false;
-}
-
-bool FullReader::execute(ISsdApp* app, LBA lba, DATA data)
-{
-	for (LBA lba = 0; lba < SSD_MAX_SIZE; ++lba)
-	{
-		if (!OuputDecoratedReader::execute(app, lba, data)) return false;
-	}
-	print("[Full Read] Done");
-	return true;
-}
-
-bool Comparer::execute(ISsdApp* app, LBA lba, DATA data)
-{
-	if (!Reader::execute(app, lba, data)) return false;
-	if (!Compare(data, Reader::GetResultFromFile())) return false;
-	return true;
-}
-
-bool Comparer::Compare(const DATA expectedData, const string &readResult)
-{
-	DATA readData = TestShellUtils::stringToUnsignedInt(readResult);
-	if (expectedData == readData) return true;
-
-	SHELL_LOG("[Compare] Fail Expected ", expectedData);
-	SHELL_LOG("[Compare] Fail Real ", readData);
-
-	return false;
 }
 
 bool Helper::IsValidCommand(const vector<string>& tokens) {
@@ -277,74 +81,6 @@ bool Eraser::ExecuteChunkedErase(ISsdApp* app, LBA startLba, SIZE size)
 	return true;
 }
 
-bool SizeEraser::IsValidCommand(const vector<string>& tokens)
-{
-	if (tokens[CMD_IDX] == CMD && tokens.size() == NEEDED_TOKEN_COUNT) return true;
-	dbgPrint("Command parameter is not enough! needed param count:" + to_string(NEEDED_TOKEN_COUNT) + ", current param count:" + to_string(tokens.size()));
-	return false;
-}
-
-bool SizeEraser::execute(ISsdApp* app, LBA startLba, SIZE size)
-{
-	if (size == 0) return true;
-
-	pair<LBA, SIZE> eraseRange = CalculateStartLbaAndSize(startLba, size);
-
-	if (!Eraser::execute(app, eraseRange.first, eraseRange.second))
-	{
-		print("[Erase] Fail");
-		return false;
-	}
-
-	print("[Erase] Done");
-	return true;
-}
-
-pair<LBA, SIZE> SizeEraser::CalculateStartLbaAndSize(LBA lba, SIZE size)
-{
-	int realSize = static_cast<int>(size);
-	if (realSize >= 0) return { lba, size };
-
-	int savedStartLba = static_cast<int>(lba);
-	LBA newStartLba = static_cast<LBA>(max(savedStartLba + realSize + 1, 0));
-	SIZE newSize = static_cast<SIZE>(savedStartLba - newStartLba + 1);
-
-	return { newStartLba, newSize };
-}
-
-bool RangeEraser::IsValidCommand(const vector<string>& tokens)
-{
-	if (tokens[CMD_IDX] == CMD && tokens.size() == NEEDED_TOKEN_COUNT) return true;
-	dbgPrint("Command parameter is not enough! needed param count:" + to_string(NEEDED_TOKEN_COUNT) + ", current param count:" + to_string(tokens.size()));
-	return false;
-}
-
-bool RangeEraser::execute(ISsdApp* app, LBA startLba, LBA endLba)
-{
-	int realEndLba = static_cast<int>(endLba);
-	if (realEndLba < 0) return false;
-
-	pair<LBA, SIZE> eraseRange = GetOrderedLbaRange(startLba, endLba);
-
-	if (!Eraser::execute(app, eraseRange.first, eraseRange.second))
-	{
-		print("[Erase Range] Fail");
-		return false;
-	}
-
-	print("[Erase Range] Done");
-	return true;
-}
-
-pair<LBA, SIZE> RangeEraser::GetOrderedLbaRange(LBA startLba, LBA endLba)
-{
-	LBA changedStartLba, changedEndLba;
-	std::tie(changedStartLba, changedEndLba) = std::minmax(startLba, endLba);
-
-	SIZE size = changedEndLba - changedStartLba + 1;
-
-	return { changedStartLba, size };
-}
 
 bool Flusher::IsValidCommand(const vector<string>& tokens) {
 	if (tokens[CMD_IDX] == CMD && tokens.size() == NEEDED_TOKEN_COUNT) return true;
